Skip existing files

+++ src/main.rs
@@ -5,6 +5,8 @@ use std::{
 
 use anyhow::{Context, Ok, Result, bail};
 use clap::Parser;
+use sha2::{Digest, Sha256};
+use std::io::Read;
 use walkdir::WalkDir;
 
 #[derive(Parser, Debug)]
@@ -24,16 +26,24 @@ struct Args {
 
     #[arg(long)]
     delete: bool,
+
+    #[arg(short, long)]
+    checksum: bool,
 }
+
 
 fn main() -> Result<()> {
     let args = Args::parse();
+
+    if !args.source.exists() {
+        bail!("Invalid Source path: {:?}", args.source);
+    }
 
     if args.dry_run {
         println!("DRY RUN MODE - No files will be modified")
     }
     if args.source.is_file() {
-        sync_file(&args.source, &args.destination)?;
+        sync_file(&args, &args.source, &args.destination)?;
     } else if args.source.is_dir() {
         if !args.recursive {
             bail!("Source is a directory. Use -r / --recursive to sync directories")
@@ -44,8 +54,24 @@ fn main() -> Result<()> {
     Ok(())
 }
 
-fn sync_file(source: &Path, destination: &Path) -> Result<()> {
-    fs::copy(source, destination)?;
+fn sync_file(args: &Args, source: &Path, destination: &Path) -> Result<()> {
+    if let Some(parent) = destination.parent() {
+        if !args.dry_run && !parent.exists() {
+            fs::create_dir_all(parent)?;
+        }
+    }
+
+    let should_copy = if !destination.exists() {
+        true
+    } else if args.checksum {
+        !checksums_match(source, destination)?
+    } else {
+        !metadata_match(source, destination)?
+    };
+
+    if should_copy && !args.dry_run {
+        fs::copy(source, destination)?;
+    }
 
     Ok(())
 }
@@ -66,7 +92,7 @@ fn sync_directory(args: &Args) -> Result<()> {
         let relative_path = source_file.strip_prefix(&args.source)?;
         let dest_file = args.destination.join(relative_path);
 
-        sync_file(source_file, &dest_file)?;
+        sync_file(args, source_file, &dest_file)?;
     }
     if args.delete && args.destination.exists() {
         delete_extra_files(args)?;
@@ -100,4 +126,33 @@ fn delete_extra_files(args: &Args) -> Result<()> {
     }
 
     Ok(())
+}
+
+fn metadata_match(source: &Path, destination: &Path) -> Result<bool> {
+    let source_meta = fs::metadata(source)?;
+    let dest_meta = fs::metadata(destination)?;
+
+    Ok(source_meta.len() == dest_meta.len() && source_meta.modified()? <= dest_meta.modified()?)
+}
+
+fn checksums_match(source: &Path, destination: &Path) -> Result<bool> {
+    let source_hash = calculate_checksum(source)?;
+    let dest_hash = calculate_checksum(destination)?;
+    Ok(source_hash == dest_hash)
+}
+
+fn calculate_checksum(path: &Path) -> Result<Vec<u8>> {
+    let mut file = fs::File::open(path)?;
+    let mut hasher = Sha256::new();
+    let mut buffer = [0; 8192];
+
+    loop {
+        let bytes_read = file.read(&mut buffer)?;
+        if bytes_read == 0 {
+            break;
+        }
+        hasher.update(&buffer[..bytes_read]);
+    }
+
+    Ok(hasher.finalize().to_vec())
 }
